# src/routes/propertiesRouter.ts
import { Request, Router } from "express";
import propertiesService from "../services/propertiesService";
import { cloudinary, upload } from "../config/cloudinary";
import authService from "../services/authService";
import { getPublicIdFromUrl } from "../utils";
import {
  objectIdSchema,
  pageParamsSchema,
  propertySchema,
  searchParamsSchema,
} from "../schemas/propertySchemas";
import z from "zod";

type PropertyFormBody = Record<string, unknown>;

const propertiesRouter = Router();

propertiesRouter.get("/", async (req, res, next) => {
  try {
    const parsed = pageParamsSchema.safeParse(req.query);

    if (!parsed.success) {
      res.status(400).json({ error: "Invalid query parameters" });
      return;
    }

    const { properties, total } = await propertiesService.getProperties(
      parsed.data,
    );
    res.set("X-Total-Count", String(total));
    res.json(properties);
  } catch (error) {
    next(error);
  }
});

propertiesRouter.get("/search", async (req, res, next) => {
  try {
    const parsed = searchParamsSchema.safeParse(req.query);

    if (!parsed.success) {
      res.status(400).json({ error: "Invalid query parameters" });
      return;
    }

    const pageParams = pageParamsSchema.parse(req.query);

    const result = await propertiesService.searchProperties(
      parsed.data,
      pageParams,
    );
    res.set("X-Total-Count", String(result.total));
    res.json(result.properties);
  } catch (error) {
    next(error);
  }
});

const deleteUploadedImages = (imageUrls: string[]) => {
  const deletePromises = imageUrls.map((imageUrl) => {
    const publicId = getPublicIdFromUrl(imageUrl);
    return cloudinary.uploader.destroy(publicId);
  });
  return Promise.all(deletePromises);
};

propertiesRouter.post(
  "/",
  authService.authenticate,
  upload.array("images", 5),
  async (req, res, next) => {
    const imageUrls =
      (req.files as Express.Multer.File[])?.map((file) => file.path) || [];

    if (imageUrls.length === 0) {
      res.status(400).json({ error: "At least one image is required" });
      return;
    }

    try {
      const body = req.body as PropertyFormBody;

      const propertyData = propertySchema.parse({
        title: body.title,
        ref: body.ref,
        description: body.description,
        price: Number(body.price),
        propertyType: body.propertyType,
        location: body.location,
        area: Number(body.area),
        status: body.status,
        images: imageUrls,
        bedrooms: body.bedrooms ? Number(body.bedrooms) : undefined,
        bathrooms: body.bathrooms ? Number(body.bathrooms) : undefined,
        floor: body.floor ? Number(body.floor) : undefined,
        parking: body.parking === "true",
      });

      const newProperty = await propertiesService.addProperty(propertyData);
      res.status(201).json(newProperty);
    } catch (error) {
      await deleteUploadedImages(imageUrls);
      if (error instanceof z.ZodError) {
        res.status(400).send({ error: error.issues });
        return;
      }
      next(error);
    }
  },
);
propertiesRouter.get("/:id", async (req, res, next): Promise<void> => {
  if (!objectIdSchema.safeParse(req.params.id).success) {
    res.status(400).json({ error: "Invalid property id" });
    return;
  }

  try {
    const property = await propertiesService.getProperty(req.params.id);
    if (!property) {
      res.status(404).json({ error: "Property not found" });
      return;
    }
    res.json(property);
  } catch (error) {
    next(error);
  }
});

propertiesRouter.delete(
  "/:id",
  authService.authenticate,
  async (req: Request<{ id: string }>, res, next) => {
    if (!objectIdSchema.safeParse(req.params.id).success) {
      res.status(400).json({ error: "Invalid property id" });
      return;
    }

    try {
      const property = await propertiesService.getProperty(req.params.id);

      if (!property) {
        res.status(404).json({ error: "Property not found" });
        return;
      }

      const deleteImagePromises = property.images.map((imageUrl) => {
        const publicId = getPublicIdFromUrl(imageUrl);
        return cloudinary.uploader.destroy(publicId);
      });

      await Promise.all(deleteImagePromises);
      const deletedProperty = await propertiesService.deleteProperty(
        req.params.id,
      );

      res.json({
        message: "Property and images deleted successfully",
        deletedProperty,
      });
    } catch (error) {
      next(error);
    }
  },
);

propertiesRouter.put(
  "/:id",
  authService.authenticate,
  upload.array("images", 5),
  async (req: Request<{ id: string }>, res, next) => {
    if (!objectIdSchema.safeParse(req.params.id).success) {
      res.status(400).json({ error: "Invalid property id" });
      return;
    }

    const newImageUrls =
      (req.files as Express.Multer.File[])?.map((file) => file.path) || [];

    try {
      // Get existing property
      const existingProperty = await propertiesService.getProperty(
        req.params.id,
      );

      if (!existingProperty) {
        await deleteUploadedImages(newImageUrls);
        res.status(404).json({ error: "Property not found" });
        return;
      }

      // Parse existing images from request body (if frontend sends them)
      const rawExistingImages = (req.body as PropertyFormBody).existingImages;
      let existingImages: string[];
      if (typeof rawExistingImages === "string") {
        try {
          existingImages = JSON.parse(rawExistingImages) as string[];
        } catch {
          await deleteUploadedImages(newImageUrls);
          res.status(400).json({ error: "Invalid existingImages" });
          return;
        }
      } else {
        existingImages = existingProperty.images;
      }

      // Combine existing + new images
      const allImages = [...existingImages, ...newImageUrls];

      // Validate max 5 images
      if (allImages.length > 5) {
        await deleteUploadedImages(newImageUrls);
        res.status(400).json({ error: "Maximum 5 images allowed" });
        return;
      }

      const body = req.body as PropertyFormBody;

      const propertyData = propertySchema.parse({
        title: body.title,
        ref: body.ref,
        description: body.description,
        price: Number(body.price),
        propertyType: body.propertyType,
        location: body.location,
        area: Number(body.area),
        status: body.status,
        images: allImages,
        bedrooms: body.bedrooms ? Number(body.bedrooms) : undefined,
        bathrooms: body.bathrooms ? Number(body.bathrooms) : undefined,
        floor: body.floor ? Number(body.floor) : undefined,
        parking: body.parking === "true",
      });

      // Find deleted images (were in DB but not in request)
      const deletedImages = existingProperty.images.filter(
        (img) => !existingImages.includes(img),
      );

      // Delete removed images from Cloudinary
      if (deletedImages.length > 0) {
        await deleteUploadedImages(deletedImages);
      }

      const updatedProperty = await propertiesService.updateProperty(
        req.params.id,
        propertyData,
      );

      res.json(updatedProperty);
    } catch (error) {
      await deleteUploadedImages(newImageUrls);
      if (error instanceof z.ZodError) {
        res.status(400).json({ error: error.issues });
        return;
      }
      next(error);
    }
  },
);

export default propertiesRouter;